#189 [KeyLoader] Fix tests on HHVM using file_get_contents
Tests are broken on HHVM since the KeyLoader uses `file://`, the problem was hidden before that the OpenSSLKeyLoader exists, because this part was not precisely tested).

So I did the same as in the new key loader added in `2.0`. It now uses `file_get_contents([KEY PATH])` rather than `file://[KEY PATH]` as first argument of `open_ssl_pkey_get`.