Make openssl key loader service deprecated
Service OpenSSLKeyLoader is deprecated since version 2.5. It triggers
deprecation error as soon as someone tries to include it. Symfony
framework tries to create reflection of it while building the container
and it triggers deprecation as well. Symfony doesn't know anything about this
service being deprecated. Service should be declared as deprecated in
service config, then Symfony won't touch it.

[https://github.com/lexik/LexikJWTAuthenticationBundle/issues/523]