Fixes TypeError in JWTManager
When the TokenExtractor extracts a token, it will return
`string|false`.  The JWTAuthenticator then passes this value to
`JWTManager::doAuthenticate` which can only accept a string.  If a false
is returned by the TokenExtractor, PHP throws a type error.  This commit
checks the return value and throws an exception if it returns false.

Issue: #1066